+++ src/configValidation.ts
@@ -1,0 +1,10 @@
+import { ConfigWithDefaults } from "./config";
+
+export function validate(config: ConfigWithDefaults) {
+  if ((config.tlsCertificateFile && !config.tlsCertificateKeyFile) ||
+       (!config.tlsCertificateFile && config.tlsCertificateKeyFile)) {
+    throw new Error(
+      'tlsCertificateFile and tlsCertificateKeyFile must both be set!'
+    );
+  }
+}